# Remove some Expr clones in `EliminateCrossJoin`(3%-5% faster planning)